src/main/java/com/leetcode/p320/OddEvenLinkedList328.java
package com.leetcode.p320;

import com.leetcode.ListNode;

/**
 * Created by sunilpatil on 5/25/17.
 */
public class OddEvenLinkedList328 {
    public ListNode oddEvenList(ListNode head) {
        if (head != null) {
            ListNode odd = head;
            ListNode even = head.next;
            ListNode evenHead = even;

            while (even != null && even.next != null) {
                odd.next = odd.next.next;
                even.next = even.next.next;
                odd = odd.next;
                even = even.next;
            }
            odd.next = evenHead;
        }
        return head;
    }
}
